+++ web/src/main/java/com/microsoft/applicationinsights/web/extensibility/initializers/WebOperationIdTelemetryInitializer.java
@@ -27,7 +27,7 @@ import com.microsoft.applicationinsights.telemetry.RequestTelemetry;
 import com.microsoft.applicationinsights.telemetry.Telemetry;
 import com.microsoft.applicationinsights.web.internal.RequestTelemetryContext;
 import com.microsoft.applicationinsights.web.internal.ThreadContext;
-import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Created by yonisha on 2/16/2015.
@@ -68,7 +68,7 @@ public class WebOperationIdTelemetryInitializer extends WebTelemetryInitializerB
         }
 
         // add correlation context to properties
-        HashMap<String, String> correlationContextMap = telemetryContext.getCorrelationContext().getMappings();
+        Map<String, String> correlationContextMap = telemetryContext.getCorrelationContext().getMappings();
         for (String key : correlationContextMap.keySet()) {
             telemetry.getProperties().putIfAbsent(key, correlationContextMap.get(key));
         }